requirements.txt
dotenv==0.0.5
Faker==24.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0